+++ Netcup.Ccp.Tests/NetcupApiClientTest.cs
@@ -5,6 +5,7 @@ using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Reflection;
+using System.Runtime.Serialization;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -94,7 +95,7 @@ namespace Netcup.Ccp.Tests
 				Message = "Login successful",
 				ServerRequestId = "Server-Request-ID",
 				StatusCode = 2000,
-				Status = "success",
+				Status = ResponseStatus.Success,
 				ResponseData = new SessionObject
 				{
 					ApiSessionId = API_SESSION_ID
@@ -206,7 +207,7 @@ namespace Netcup.Ccp.Tests
 				Message = "Logout successful",
 				ServerRequestId = "Server-Request-ID",
 				StatusCode = 2000,
-				Status = "success",
+				Status = ResponseStatus.Success,
 				ResponseData = ""
 			};
 			if (!string.IsNullOrEmpty(clientRequestId))
@@ -261,6 +262,7 @@ namespace Netcup.Ccp.Tests
 		[DataRow(null)]
 		[DataRow("")]
 		[DataRow("   ")]
+		[ExpectedException(typeof(ArgumentNullException))]
 		public async Task ShouldThrowArgumentNullExceptionForMissingApiSessionId(string id)
 		{
 			// Arrange
@@ -290,16 +292,11 @@ namespace Netcup.Ccp.Tests
 				Message = "DNS zone found",
 				ServerRequestId = "Server-Request-ID",
 				StatusCode = 2000,
-				Status = "success",
-				ResponseData = new DnsZone
+				Status = ResponseStatus.Success,
+				ResponseData = new DnsZone(ttl: 86400, refresh: 28800, retry: 7200, expire: 1209600, dnssecstatus: false)
 				{
-					DnsSecStatus = false,
-					Expire = 1209600,
 					Name = "netcup.de",
-					Refresh = 28800,
-					Retry = 7200,
 					Serial = 2024060199,
-					Ttl = 86400
 				}
 			};
 			if (!string.IsNullOrEmpty(clientRequestId))
@@ -400,24 +397,18 @@ namespace Netcup.Ccp.Tests
 				Message = "DNS records found",
 				ServerRequestId = "Server-Request-ID",
 				StatusCode = 2000,
-				Status = "success",
+				Status = ResponseStatus.Success,
 				ResponseData = new DnsRecordSet
 				{
 					Records =
 					[
-						new DnsRecord
+						new DnsRecord("@", DnsRecordType.A, "46.38.224.30")
 						{
 							Id = 123,
-							Hostname = "@",
-							Type = "A",
-							Destination = "46.38.224.30"
 						},
-						new DnsRecord
+						new DnsRecord("@", DnsRecordType.Aaaa, "2a03:4000::e01e")
 						{
 							Id = 456,
-							Hostname = "@",
-							Type = "AAAA",
-							Destination = "2a03:4000::e01e"
 						}
 					]
 				}
@@ -456,12 +447,12 @@ namespace Netcup.Ccp.Tests
 
 			Assert.AreEqual((uint?)123, response.ResponseData.Records[0].Id);
 			Assert.AreEqual("@", response.ResponseData.Records[0].Hostname);
-			Assert.AreEqual("A", response.ResponseData.Records[0].Type);
+			Assert.AreEqual(DnsRecordType.A, response.ResponseData.Records[0].Type);
 			Assert.AreEqual("46.38.224.30", response.ResponseData.Records[0].Destination);
 
 			Assert.AreEqual((uint?)456, response.ResponseData.Records[1].Id);
 			Assert.AreEqual("@", response.ResponseData.Records[1].Hostname);
-			Assert.AreEqual("AAAA", response.ResponseData.Records[1].Type);
+			Assert.AreEqual(DnsRecordType.Aaaa, response.ResponseData.Records[1].Type);
 			Assert.AreEqual("2a03:4000::e01e", response.ResponseData.Records[1].Destination);
 
 			Assert.AreEqual(1, _httpRequestMessageCallbacks.Count);
@@ -498,16 +489,11 @@ namespace Netcup.Ccp.Tests
 				Message = "DNS zone successful updated",
 				ServerRequestId = "Server-Request-ID",
 				StatusCode = 2000,
-				Status = "success",
-				ResponseData = new DnsZone
+				Status = ResponseStatus.Success,
+				ResponseData = new DnsZone(ttl: 86400, refresh: 28800, retry: 7200, expire: 1209600, dnssecstatus: true)
 				{
 					Name = "netcup.de",
-					Ttl = 86400,
 					Serial = 2024060199,
-					Refresh = 28800,
-					Retry = 7200,
-					Expire = 1209600,
-					DnsSecStatus = true,
 				}
 			};
 			if (!string.IsNullOrEmpty(clientRequestId))
@@ -596,13 +582,7 @@ namespace Netcup.Ccp.Tests
 		public async Task ShouldThrowOutOfRangeForExpireOnUpdateDnsZone()
 		{
 			// Arrange
-			var zone = new DnsZone
-			{
-				Expire = 0,
-				Refresh = 1,
-				Retry = 1,
-				Ttl = 1
-			};
+			var zone = new DnsZone(ttl: 1, refresh: 1, retry: 1, expire: 0, dnssecstatus: false);
 			var client = GetClient();
 
 			// Act
@@ -616,13 +596,7 @@ namespace Netcup.Ccp.Tests
 		public async Task ShouldThrowOutOfRangeForRefreshOnUpdateDnsZone()
 		{
 			// Arrange
-			var zone = new DnsZone
-			{
-				Expire = 1,
-				Refresh = 0,
-				Retry = 1,
-				Ttl = 1
-			};
+			var zone = new DnsZone(ttl: 1, refresh: 0, retry: 1, expire: 1, dnssecstatus: false);
 			var client = GetClient();
 
 			// Act
@@ -636,13 +610,7 @@ namespace Netcup.Ccp.Tests
 		public async Task ShouldThrowOutOfRangeForRetryOnUpdateDnsZone()
 		{
 			// Arrange
-			var zone = new DnsZone
-			{
-				Expire = 1,
-				Refresh = 1,
-				Retry = 0,
-				Ttl = 1
-			};
+			var zone = new DnsZone(ttl: 1, refresh: 1, retry: 0, expire: 1, dnssecstatus: false);
 			var client = GetClient();
 
 			// Act
@@ -656,13 +624,7 @@ namespace Netcup.Ccp.Tests
 		public async Task ShouldThrowOutOfRangeForTtlOnUpdateDnsZone()
 		{
 			// Arrange
-			var zone = new DnsZone
-			{
-				Expire = 1,
-				Refresh = 1,
-				Retry = 1,
-				Ttl = 0
-			};
+			var zone = new DnsZone(ttl: 0, refresh: 1, retry: 1, expire: 1, dnssecstatus: false);
 			var client = GetClient();
 
 			// Act
@@ -672,12 +634,12 @@ namespace Netcup.Ccp.Tests
 		}
 
 		[DataTestMethod]
-		[DataRow(null, "A", null, null, null)]
-		[DataRow(111u, "A", null, null, null)]
-		[DataRow(null, "MX", 10u, null, null)]
-		[DataRow(222u, "A", null, true, null)]
-		[DataRow(null, "A", null, null, "ClientRequestId")]
-		public async Task ShouldUpdateDnsRecords(uint? id, string type, uint? prio, bool? delete, string clientRequestId)
+		[DataRow(null, DnsRecordType.A, null, null, null)]
+		[DataRow(111u, DnsRecordType.Aaaa, null, null, null)]
+		[DataRow(null, DnsRecordType.Mx, 10u, null, null)]
+		[DataRow(222u, DnsRecordType.A, null, true, null)]
+		[DataRow(null, DnsRecordType.Ns, null, null, "ClientRequestId")]
+		public async Task ShouldUpdateDnsRecords(uint? id, DnsRecordType type, uint? prio, bool? delete, string clientRequestId)
 		{
 			// Arrange
 			var res = new ResponseMessage<DnsRecordSet>
@@ -688,24 +650,18 @@ namespace Netcup.Ccp.Tests
 				Message = "DNS records successful updated",
 				ServerRequestId = "Server-Request-ID",
 				StatusCode = 2000,
-				Status = "success",
+				Status = ResponseStatus.Success,
 				ResponseData = new DnsRecordSet
 				{
 					Records =
 					[
-						new DnsRecord
+						new DnsRecord("@", DnsRecordType.A, "46.38.224.30")
 						{
 							Id = 123,
-							Hostname = "@",
-							Type = "A",
-							Destination = "46.38.224.30"
 						},
-						new DnsRecord
+						new DnsRecord("@", DnsRecordType.Aaaa, "2a03:4000::e01e")
 						{
 							Id = 456,
-							Hostname = "@",
-							Type = "AAAA",
-							Destination = "2a03:4000::e01e"
 						}
 					]
 				}
@@ -717,13 +673,10 @@ namespace Netcup.Ccp.Tests
 			{
 				Records =
 				[
-					new DnsRecord
+					new DnsRecord("@", type, "46.38.224.30")
 					{
 						Id = id,
-						Hostname = "@",
-						Type = type,
 						Priority = prio,
-						Destination = "46.38.224.30",
 						DeleteRecord = delete,
 						State = "unknown"
 					}
@@ -761,12 +714,12 @@ namespace Netcup.Ccp.Tests
 
 			Assert.AreEqual((uint?)123, response.ResponseData.Records[0].Id);
 			Assert.AreEqual("@", response.ResponseData.Records[0].Hostname);
-			Assert.AreEqual("A", response.ResponseData.Records[0].Type);
+			Assert.AreEqual(DnsRecordType.A, response.ResponseData.Records[0].Type);
 			Assert.AreEqual("46.38.224.30", response.ResponseData.Records[0].Destination);
 
 			Assert.AreEqual((uint?)456, response.ResponseData.Records[1].Id);
 			Assert.AreEqual("@", response.ResponseData.Records[1].Hostname);
-			Assert.AreEqual("AAAA", response.ResponseData.Records[1].Type);
+			Assert.AreEqual(DnsRecordType.Aaaa, response.ResponseData.Records[1].Type);
 			Assert.AreEqual("2a03:4000::e01e", response.ResponseData.Records[1].Destination);
 
 			Assert.AreEqual(1, _httpRequestMessageCallbacks.Count);
@@ -793,7 +746,7 @@ namespace Netcup.Ccp.Tests
 
 			Assert.AreEqual(1, req["param"]["dnsrecordset"].Value<JArray>("dnsrecords").Count);
 			Assert.AreEqual("@", req["param"]["dnsrecordset"]["dnsrecords"][0].Value<string>("hostname"));
-			Assert.AreEqual(type, req["param"]["dnsrecordset"]["dnsrecords"][0].Value<string>("type"));
+			Assert.AreEqual(GetEnumMemberValue(type), req["param"]["dnsrecordset"]["dnsrecords"][0].Value<string>("type"));
 			Assert.AreEqual("46.38.224.30", req["param"]["dnsrecordset"]["dnsrecords"][0].Value<string>("destination"));
 			Assert.IsFalse(req["param"]["dnsrecordset"]["dnsrecords"].Value<JObject>(0).ContainsKey("state"));
 
@@ -835,12 +788,9 @@ namespace Netcup.Ccp.Tests
 			{
 				Records =
 				[
-					new DnsRecord
+					new DnsRecord("@", DnsRecordType.A, "46.38.224.30")
 					{
 						Id = 0,
-						Hostname = "@",
-						Type= "A",
-						Destination = "46.38.224.30",
 					}
 				]
 			};
@@ -864,12 +814,9 @@ namespace Netcup.Ccp.Tests
 			{
 				Records =
 				[
-					new DnsRecord
+					new DnsRecord(hostname, DnsRecordType.A, "46.38.224.30")
 					{
 						Id = null,
-						Hostname = hostname,
-						Type= "A",
-						Destination = "46.38.224.30",
 					}
 				]
 			};
@@ -890,12 +837,9 @@ namespace Netcup.Ccp.Tests
 			{
 				Records =
 				[
-					new DnsRecord
+					new DnsRecord("@", 0, "46.38.224.30")
 					{
 						Id = null,
-						Hostname = "@",
-						Type= "AA",
-						Destination = "46.38.224.30",
 					}
 				]
 			};
@@ -916,13 +860,10 @@ namespace Netcup.Ccp.Tests
 			{
 				Records =
 				[
-					new DnsRecord
+					new DnsRecord("@", DnsRecordType.Mx, "46.38.224.30")
 					{
 						Id = null,
-						Hostname = "@",
-						Type= "MX",
 						Priority = null,
-						Destination = "46.38.224.30",
 					}
 				]
 			};
@@ -946,12 +887,9 @@ namespace Netcup.Ccp.Tests
 			{
 				Records =
 				[
-					new DnsRecord
+					new DnsRecord("@", DnsRecordType.A, destination)
 					{
 						Id = null,
-						Hostname = "@",
-						Type= "A",
-						Destination = destination,
 					}
 				]
 			};
@@ -972,12 +910,9 @@ namespace Netcup.Ccp.Tests
 			{
 				Records =
 				[
-					new DnsRecord
+					new DnsRecord("@", DnsRecordType.A, "46.38.224.30")
 					{
 						Id = null,
-						Hostname = "@",
-						Type= "A",
-						Destination = "46.38.224.30",
 						DeleteRecord = true
 					}
 				]
@@ -991,6 +926,17 @@ namespace Netcup.Ccp.Tests
 		}
 
 		#endregion DNS API
+
+		private string GetEnumMemberValue(Enum value)
+		{
+			var fi = value.GetType().GetField(value.ToString());
+
+			var attr = fi.GetCustomAttributes(typeof(EnumMemberAttribute), inherit: false)
+				.Cast<EnumMemberAttribute>()
+				.FirstOrDefault();
+
+			return attr.Value;
+		}
 
 		private NetcupApiClient GetClient()
 		{