+++ Netcup.Ccp/Models/DnsZone.cs
@@ -1,16 +1,16 @@
-﻿namespace AMWD.Net.Api.Netcup.Ccp.Models
+﻿namespace AMWD.Net.Api.Netcup.Ccp
 {
 	/// <summary>
 	/// DNS zone.
 	/// </summary>
 	[System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-	public class DnsZone
+	public class DnsZone(uint ttl, uint refresh, uint retry, uint expire, bool dnssecstatus)
 	{
 		/// <summary>
 		/// Name of the zone - this is a domain name.
 		/// </summary>
 		[JsonProperty("name", Required = Required.AllowNull)]
-		public string Name { get; set; }
+		public string? Name { get; set; }
 
 		/// <summary>
 		/// time-to-live Time in seconds a domain name is cached locally before expiration and return to authoritative nameservers for updated information.
@@ -18,7 +18,7 @@
 		/// Recommendation: 3600 to 172800
 		/// </summary>
 		[JsonProperty("ttl", Required = Required.Always)]
-		public uint Ttl { get; set; }
+		public uint Ttl { get; set; } = ttl;
 
 		/// <summary>
 		/// Serial of zone.
@@ -33,7 +33,7 @@
 		/// Recommendation: 3600 to 14400
 		/// </summary>
 		[JsonProperty("refresh", Required = Required.Always)]
-		public uint Refresh { get; set; }
+		public uint Refresh { get; set; } = refresh;
 
 		/// <summary>
 		/// Time in seconds primary name server waits if an attempt to refresh by a secondary name server failed.
@@ -41,7 +41,7 @@
 		/// Recommendation: 900 to 3600
 		/// </summary>
 		[JsonProperty("retry", Required = Required.Always)]
-		public uint Retry { get; set; }
+		public uint Retry { get; set; } = retry;
 
 		/// <summary>
 		/// Time in seconds a secondary name server will hold a zone before it is no longer considered authoritative.
@@ -49,14 +49,14 @@
 		/// Recommendation: 592200 to 1776600
 		/// </summary>
 		[JsonProperty("expire", Required = Required.Always)]
-		public uint Expire { get; set; }
+		public uint Expire { get; set; } = expire;
 
 		/// <summary>
 		/// Status of DNSSSEC in this nameserver.
 		/// Enabling DNSSEC possible every 24 hours.
 		/// </summary>
 		[JsonProperty("dnssecstatus", Required = Required.Always)]
-		public bool DnsSecStatus { get; set; }
+		public bool DnsSecStatus { get; set; } = dnssecstatus;
 
 		/// <inheritdoc />
 		public override string ToString()