+++ Netcup.Ccp/Models/ResponseMessage.cs
@@ -1,4 +1,4 @@
-﻿namespace AMWD.Net.Api.Netcup.Ccp.Models
+﻿namespace AMWD.Net.Api.Netcup.Ccp
 {
 	/// <summary>
 	/// Response message of a request send to the api.
@@ -10,43 +10,43 @@
 		/// Unique ID for the request, created by the server.
 		/// </summary>
 		[JsonProperty("serverrequestid", Required = Required.Always)]
-		public string ServerRequestId { get; set; }
+		public string? ServerRequestId { get; set; }
 
 		/// <summary>
 		/// Unique ID for the request, created by the client.
 		/// </summary>
 		[JsonProperty("clientrequestid")]
-		public string ClientRequestId { get; set; }
+		public string? ClientRequestId { get; set; }
 
 		/// <summary>
 		/// Name of the function that was called.
 		/// </summary>
 		[JsonProperty("action", Required = Required.Always)]
-		public string Action { get; set; }
+		public string? Action { get; set; }
 
 		/// <summary>
 		/// Staus of the Message like "error", "started", "pending", "warning" or "success".
 		/// </summary>
 		[JsonProperty("status", Required = Required.Always)]
-		public string Status { get; set; }
+		public ResponseStatus? Status { get; set; }
 
 		/// <summary>
 		/// Staus code of the Message like 2011.
 		/// </summary>
 		[JsonProperty("statuscode", Required = Required.Always)]
-		public uint StatusCode { get; set; }
+		public uint? StatusCode { get; set; }
 
 		/// <summary>
 		/// Short message with information about the processing of the messsage.
 		/// </summary>
 		[JsonProperty("shortmessage", Required = Required.Always)]
-		public string Message { get; set; }
+		public string? Message { get; set; }
 
 		/// <summary>
 		/// Long message with information about the processing of the messsage.
 		/// </summary>
 		[JsonProperty("longmessage")]
-		public string Description { get; set; }
+		public string? Description { get; set; }
 	}
 
 	/// <summary>
@@ -59,7 +59,7 @@
 		/// Data from the response like domain object.
 		/// </summary>
 		[JsonProperty("responsedata")]
-		public object ResponseData { get; set; }
+		public object? ResponseData { get; set; }
 	}
 
 	/// <summary>
@@ -72,6 +72,6 @@
 		/// Data from the response like domain object.
 		/// </summary>
 		[JsonProperty("responsedata")]
-		public T ResponseData { get; set; }
+		public T? ResponseData { get; set; }
 	}
 }